\name{MKLE-package}
\alias{MKLE-package}
\alias{MKLE}
\docType{package}
\title{
Maximum kernel likelihood estimation
}
\description{
Computes the kernel density estimator for mu
}
\details{
\tabular{ll}{
Package: \tab MKLE\cr
Type: \tab Package\cr
Version: \tab 0.02\cr
Date: \tab 2006-07-12\cr
License: \tab GNU \cr
}
}
\author{
Thomas Jaki

Maintainer: Thomas Jaki <jaki.thomas@gmail.com>
}
\references{
Silverman, B. W. (1986) Density Estimation for Statistics and Data Analysis. \emph{Chapman & Hall}\\
~ more to come ~
}
\keyword{misc}
